reorginised middleware and tools to be more organised
