+++ package.json
@@ -43,6 +43,7 @@
     "jsonwebtoken": "^8.5.1",
     "knex": "^0.20.10",
     "knex-cleaner": "^1.3.0",
+    "morgan": "^1.9.1",
     "pg": "^7.18.2",
     "stripe": "^8.32.0",
     "supertest": "^4.0.2"
