+++ server/server.js
@@ -1,19 +1,21 @@
-const express = require('express'),
-  configMiddleware = require('../middleware'),
-  server = express(),
-  loginRouter = require('./auth/login'),
-  logoutRouter = require('./auth/logout'),
-  registerRouter = require('./auth/register'),
+const server = require('express')(),
+  json = require('express').json(),
+  morgan = require('morgan')('dev'),
+  helmet = require('helmet')(),
+  cors = require('cors')(),
+  loginRouter = require('./login'),
+  logoutRouter = require('./logout'),
+  registerRouter = require('./registerAdmin'),
   membersRouter = require('./members/member-router'),
-  stripeRouter = require('./stripe')
+  backgroundRouter = require('./background')
 
-configMiddleware(server)
+server.use(json, morgan, helmet, cors)
 
 server.use('/login', loginRouter)
 server.use('/logout', logoutRouter)
 server.use('/register', registerRouter)
 server.use('/members', membersRouter)
-server.use('/stripe', stripeRouter)
+server.use('/background', backgroundRouter)
 
 server.get('/', (req, res) => {
   res.status(200).json({ Server: 'Running' })
